requirements.txt
# Основные зависимости (МИНИМУМ)
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
t-tech-investments>=0.3.4
scipy>=1.11.0

# Визуализация (для анализа)
matplotlib>=3.7.0
seaborn>=0.12.0

# Уведомления
requests>=2.31.0
python-telegram-bot>=20.0
